+++ ios/PostbackBridge.m
@@ -52,14 +52,4 @@ RCT_EXTERN_METHOD(isSdkDisabled:(RCTPromiseResolveBlock)resolve
 RCT_EXTERN_METHOD(destroy:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)
 
-// Utility
-RCT_EXTERN_METHOD(getDeviceInfo:(RCTPromiseResolveBlock)resolve
-                  rejecter:(RCTPromiseRejectBlock)reject)
-
-RCT_EXTERN_METHOD(getWebViewUserAgent:(RCTPromiseResolveBlock)resolve
-                  rejecter:(RCTPromiseRejectBlock)reject)
-
-RCT_EXTERN_METHOD(getAdServicesToken:(RCTPromiseResolveBlock)resolve
-                  rejecter:(RCTPromiseRejectBlock)reject)
-
 @end