+++ lib/typescript/index.d.ts
@@ -1,4 +1,3 @@
 export { Postback } from "./Postback";
-export { NativePostback } from "./NativePostback";
-export type { PostbackConfig, PostbackOptions, EventType, EventParams, AttributionResult, InstallResponse, InstallType, DeviceInfo, LogLevel, TestEventResult, NativePostbackModule, } from "./types";
+export type { PostbackConfig, PostbackOptions, EventType, EventParams, AttributionResult, InstallResponse, LogLevel, TestEventResult, } from "./types";
 //# sourceMappingURL=index.d.ts.map